+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Added ability to handle connecting to multiple node cluster
 
 ### Changed
+- \_process_queue:  Added microseconds to the datetime group.
 - monitor_queue:  Setup heartbeat and host_list parameters for rabbitmq_class.RabbitMQCon.
 - config/rabbitmq.py.TEMPLATE:  Added heartbeat and host_list entries.
 - Removed unnecessary \*\*kwargs in function argument list.